move some functions from setup to support
