+++ src/test/K6/src/tests/app/performance/uploadanddownload.js
@@ -29,6 +29,7 @@ import * as apps from '../../../api/platform/storage/applications.js';
 import * as storageData from '../../../api/platform/storage/data.js';
 import { deleteSblInstance } from '../../../api/platform/storage/messageboxinstances.js';
 import * as setUpData from '../../../setup.js';
+import * as support from '../../../support.js';
 
 const instanceFormDataXml = open('../../../data/' + level2App + '.xml');
 const appOwner = __ENV.org;
@@ -58,7 +59,7 @@ export function setup() {
   var data = {};
   var totalIterations = options.iterations ? options.iterations : 1;
   attachmentDistribution = attachmentDistribution ? attachmentDistribution : '';
-  let attachmentTypes = setUpData.buildAttachmentTypeArray(attachmentDistribution, totalIterations);
+  let attachmentTypes = support.buildAttachmentTypeArray(attachmentDistribution, totalIterations);
   data.attachmentTypes = attachmentTypes;
   return data;
 }
